Rename AudioCodec sub classes.
